feat: Arrumando estrutura de pastas, Ajustando um 'theme global', colocando na tela principal o login para eu poder visualizar a tela e instalando dependencias
