--- front/src/screens/.gitkeep
@@ -1,1 +1,0 @@
-# Telas da aplicação
