+++ front/src/theme/colors.ts
@@ -1,0 +1,11 @@
+export const colors = {
+  darkBg: "#2C3E50",
+  sheetBg: "#F5F0EB",
+  inputBg: "#3D5166",
+  accent: "#B5845A",
+  textLight: "#FFFFFF",
+  textDark: "#2C3E50",
+  textMuted: "#9AA5B4",
+  textSubtle: "#9A8F88",
+  divider: "#D5C9BE",
+};
