+++ front/src/types/index.ts
@@ -1,0 +1,2 @@
+// Add your TypeScript types here
+// Example: export type User = { id: string; name: string; };
